Add a simple unmask app
This was derived from apps/asynchromix/asynchromix.[py,sol].

The "mixing" part, of the original asynchromix app was
removed to make it into a simple app in which a client sends
a mask message and the MPC network unmasks it.

Perhaps there is no need for an ethereum-based coordinator for such a
simple case but the idea is to provide a basis for more complex
applications such as message mixing and cross-shard secure
communications.