wip: refine the modularization of key components
The goal is to organize the code such that the pieces belonging to
different "actors" are ideally fully isolated. That is, the client code
is just client code, the MPC server/network code is just MPC server code,
the contract deployment code is just contract deployment code, etc. With
this separation it should be possible to run each component
independently of each other, such as in different processes, or different
containers.