wip: setup phase writes common config file
The setup phase for now is responsible to deploy the contract and to
provide the contract address to the MPC servers and clients.

Additionaly, the contract deployer, the MPC servers and clients, need
an eth address and for now a "dummy" eth address is assigned to each
participant. The addresses can be added to the common config file.

next:
* make sure all common data is in the config file
* update client config with just client config data
* use config for mpc network